DS-351:Refactore dialog stories
